+++ packages/lib/src/core/blueprints/Blueprints.js
@@ -254,20 +254,6 @@ class Blueprints {
         addPresetStyles(styles),
         isUsingCustomInfoElements
       ); // TODO make sure the processLayouts is up to date. delete addLayoutStyles from layoutsHelper when done with it...
-
-      const selectedLayoutVars = [
-        'galleryLayout',
-        'galleryThumbnailsAlignment',
-        'magicLayoutSeed',
-        'cubeType',
-        'isVertical',
-        'scrollDirection',
-        'enableInfiniteScroll',
-      ];
-      formattedStyles.selectedLayout = selectedLayoutVars
-        .map((key) => String(formattedStyles[key]))
-        .join('|');
-      formattedStyles.layoutsVersion = 2;
       changed = true;
     }