+++ packages/playground/src/components/SideBar/comments.js
@@ -1,9 +1,6 @@
 const comment =
 ['setting that now we save in "appSettings" storage but can be converted to styleParams \n'+
-'watermarkOpacity\n'+
-'watermarkSize\n'+
-'watermarkDock\n'
-,
+
 '--> properties to support when building settings:\n'+
 'Font \\ Color picker: \n'+
 'INPUT_TYPES.TEXT - maxLength\n'+
@@ -31,7 +28,6 @@ const comment =
 '*) I saw that mobile properties are generated with \'m_\' in settings-controller \n'+
 '*) properties that are not implemented  - mobile needs to be checked - this is not a complete list \n'+
 'm_numberOfImagesPerRow\n'+
-'mobilePanorama\n'+
 'm_isAutoSlideshow\n'+
 'm_galleryLayout\n'+
 'm_slideshowLoop - defined but I don\'t see where we use it\n'+